requirements.txt
pysnmp==4.4.12
python-nmap==0.6.1
SQLAlchemy==1.3.13
netifaces==0.10.9
pysnmp-mibs==0.1.6
reactor==0.1.4
twisted==19.10.0